Agrega componente Reunion para quitar lógica repetida entre ReuninesActivas y ReunionesCerradas
