+++ frontend/src/empezar-reunion/Reunion.js
@@ -1,0 +1,57 @@
+import React, {useEffect, useState} from "react";
+import backend from "../api/backend";
+import {CircularProgress, Paper, Table, TableBody, TableHead, TableRow} from "@material-ui/core";
+import {StyledTableCell} from "../minuta/TablaOradores.styled";
+import {TablaPinos} from "../minuta/Minuta.styled";
+import {useHistory} from "react-router-dom";
+import {ButtonIcono, ButtonReunionCerrada, SecondaryButtonReunionCerrada} from "../components/Button.styled";
+import Tooltip from "@material-ui/core/Tooltip";
+import Typography from "@material-ui/core/Typography";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
+import {faEnvelope} from "@fortawesome/free-solid-svg-icons";
+import {faSlack} from "@fortawesome/free-brands-svg-icons";
+
+const FilaReunion = ({reunion, CallToActionButton}) => {
+    return <TableRow>
+        <StyledTableCell >{reunion.temas.length > 1 ? "Reunion de Roots" : reunion.nombre || reunion.temas[0].titulo}</StyledTableCell>
+        <StyledTableCell >{reunion.temas.length > 1 ? "Roots" : reunion.temas[0].autor}</StyledTableCell>
+        <CallToActionButton/>
+    </TableRow>
+}
+
+export const Reunion = ({estaAbierta, listaDeColumnas, CallToActionButton}) => {
+
+    const history = useHistory();
+    const [reuniones,setReuniones] = useState()
+    useEffect(() => {
+        backend.obtenerReuniones(estaAbierta)
+            .then( ({reuniones}) => setReuniones(reuniones))
+    }, []);
+
+    if(!reuniones){
+        return <CircularProgress/>
+    }
+
+    return <>
+        <TablaPinos>
+            <Paper>
+                <Table aria-label="customized table">
+                    <TableHead>
+                        <TableRow>
+                            {
+                                listaDeColumnas.map((columna) => (
+                                    <StyledTableCell>{columna}</StyledTableCell>
+                                    )
+                                )
+                            }
+                        </TableRow>
+                    </TableHead>
+                    <TableBody>
+                        {reuniones.map(reunion => <FilaReunion history={history} reunion={reunion} CallToActionButton={CallToActionButton}/>)}
+                    </TableBody>
+                </Table>
+            </Paper>
+        </TablaPinos>
+    </>
+
+}
